Change c++ standard from 98 to 0x.
By changing the standard to 0x, we get to use the unique_ptr, shared_ptr
and weak_ptr templates for memory management.  The possible cost is that
these only appear after gcc 4.4 so we may not be able to compile this
everywhere.  If we run into trouble, we'll fix it.  If we don't get any
complaints, then it's OK.